beak out dynamic-struct-model.ts from metamodel.ts (using an agent)
